+++ packages/create/src/types.ts
@@ -89,6 +89,12 @@ export const AddOnBaseSchema = z.object({
     .object({
       dependencies: z.record(z.string(), z.string()).optional(),
       devDependencies: z.record(z.string(), z.string()).optional(),
+      engines: z.record(z.string(), z.string()).optional(),
+      pnpm: z
+        .object({
+          onlyBuiltDependencies: z.array(z.string()).optional(),
+        })
+        .optional(),
       scripts: z.record(z.string(), z.string()).optional(),
     })
     .optional(),
@@ -228,6 +234,7 @@ export interface Options {
   chosenAddOns: Array<AddOn>
   addOnOptions: Record<string, Record<string, any>>
   starter?: Starter | undefined
+  projectPreset?: 'default' | 'blank'
   routerOnly?: boolean
   includeExamples?: boolean
   envVarValues?: Record<string, string>