+++ cli-aliases/create-tanstack/CHANGELOG.md
@@ -1,5 +1,21 @@
 # create-tanstack
 
+## 0.54.41
+
+### Patch Changes
+
+- Updated dependencies [[`4bb1384`](https://github.com/TanStack/cli/commit/4bb13849454f5dbe126f2f67e87b0367be28bcd7)]:
+  - @tanstack/create@0.69.1
+  - @tanstack/cli@0.70.1
+
+## 0.54.40
+
+### Patch Changes
+
+- Updated dependencies [[`75db7e8`](https://github.com/TanStack/cli/commit/75db7e89a0a1125683bc75cca81acb700ef73eff), [`fb59492`](https://github.com/TanStack/cli/commit/fb59492f68aa005fbab9afad0a61aa75e90258bb)]:
+  - @tanstack/cli@0.70.0
+  - @tanstack/create@0.69.0
+
 ## 0.54.39
 
 ### Patch Changes